# frontend/src/shared/api/api.yaml
openapi: 3.0.0
paths:
...
  /games:
    get:
      operationId: GameController_getAll
      parameters: []
      responses:
        '200':
          description: Returns all game rooms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameResponseDto'
      summary: Get all game rooms
      tags:
        - Games
    post:
      operationId: GameController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameDto'
      responses:
        '201':
          description: The game room has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGameResponseDto'
        '401':
          description: Unauthorized. The user must be authenticated to create a game room.
      security:
        - bearer: []
      summary: Create a new game room
      tags:
        - Games
  /games/{id}:
    delete:
      operationId: GameController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: The unique identifier of the game room to delete.
          schema:
            example: a1b2c3d4-e5f6-7890-1234-567890abcdef
      responses:
        '204':
          description: The game room has been successfully deleted.
        '403':
          description: Forbidden. The user is not the owner of the game room.
        '404':
          description: The game room with the specified ID was not found.
      security:
        - bearer: []
      summary: Delete a game room
      tags:
        - Games
  /games/join:
    post:
      operationId: GameController_join
      parameters: []
      requestBody:
        required: true
        description: The unique identifier of the game room to join.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGameDto'
      responses:
        '200':
          description: Successfully joined the game room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponseDto'
        '403':
          description: Forbidden. The user is already in the game room.
        '404':
          description: The game room with the specified ID was not found.
      security:
        - bearer: []
      summary: Join a game room
      tags:
        - Games
  /games/code:
    post:
      operationId: GameController_getCode
      parameters: []
      requestBody:
        required: true
        description: The unique identifier of the game room to get the code.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetRoomCodeDto'
      responses:
        '200':
          description: Code was successfuly sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRoomCodeResponseDto'
        '403':
          description: Forbidden. The user is not an owner.
        '404':
          description: The game room with the specified ID was not found.
      security:
        - bearer: []
      summary: Get a room code. Only for owner
      tags:
        - Games
  /games/current:
    get:
      operationId: GameController_getMyCurrentGame
      parameters: []
      responses:
        '200':
          description: Returns roomId or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentGameResponseDto'
      security:
        - bearer: []
      summary: Get the ID of the room the user is currently in
      tags:
        - Games
  /games/validate-code:
    post:
      operationId: GameController_validateCode
      parameters: []
      requestBody:
        required: true
        description: The code and room ID to validate.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeDto'
      responses:
        '200':
          description: Returns whether the code is valid for the specified room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateCodeResponseDto'
        '404':
          description: The game room with the specified ID was not found.
      security:
        - bearer: []
      summary: Validate a game room code
      tags:
        - Games
  /history/games/{userId}:
    get:
      operationId: HistoryController_getGamesByUserId
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            default: 10
            type: number
        - name: offset
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
        - name: userId
          required: true
          in: path
          description: The ID of the user whose game history is being requested
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: List of game summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGameSummaryResponse'
      security:
        - bearer: []
      summary: Get all completed games summaries for a user
      tags:
        - history
  /history/games/{gameId}/rounds:
    get:
      operationId: HistoryController_getRoundsByGameId
      parameters:
        - name: gameId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            default: 10
            type: number
        - name: offset
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of round details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoundDetailsResponse'
      security:
        - bearer: []
      summary: Get paginated rounds for a specific game
      tags:
        - history
  /history/rounds/{roundId}:
    get:
      operationId: HistoryController_getRoundDetails
      parameters:
        - name: roundId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Round details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundDetailsResponseDto'
      security:
        - bearer: []
      summary: Get details for a specific round
      tags:
        - history
info:
  title: API Docs
  description: API Docs description
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    GameSettingsDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the game room
          example: My Awesome Game Room
        roundTimeSeconds:
          type: number
          description: The round time in seconds
          example: 60
        pointsToWin:
          type: number
          description: The points needed to win
          example: 10
        isPrivate:
          type: boolean
          description: Whether the game is private
          example: false
        level:
          type: string
          description: The difficulty level of the words used in the game
          example: medium
      required:
        - name
        - roundTimeSeconds
        - pointsToWin
        - isPrivate
        - level
    GameResponseDto:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the game
          example: 123e4567-e89b-12d3-a456-426614174000
        ownerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type: string
          enum:
            - LOBBY
            - IN_PROGRESS
            - FINISHED
        settings:
          description: Current game settings
          allOf:
            - $ref: '#/components/schemas/GameSettingsDto'
        playersCount:
          type: number
        createdAt:
          type: number
      required:
        - id
        - ownerId
        - status
        - settings
        - playersCount
        - createdAt
    CreateGameDto:
      type: object
      properties:
        name:
          type: string
          description: The name of the game
          example: My Awesome Game
        isPrivate:
          type: boolean
          description: Whether the game is private or not
          example: false
          default: false
        timeLimit:
          type: number
          description: The time limit for each turn in seconds
          example: 60
        pointsToWin:
          type: number
          description: The number of points required to win the game
          example: 30
        level:
          type: string
          description: The difficulty level of the words used in the game
          example: medium
          enum:
            - easy
            - medium
            - hard
      required:
        - name
        - isPrivate
        - timeLimit
        - pointsToWin
        - level
    CreateGameResponseDto:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the game
          example: 123e4567-e89b-12d3-a456-426614174000
        ownerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type: string
          enum:
            - LOBBY
            - IN_PROGRESS
            - FINISHED
        settings:
          description: Current game settings
          allOf:
            - $ref: '#/components/schemas/GameSettingsDto'
        playersCount:
          type: number
        createdAt:
          type: number
        code:
          type: string
      required:
        - id
        - ownerId
        - status
        - settings
        - playersCount
        - createdAt
        - code
    JoinGameDto:
      type: object
      properties:
        roomId:
          type: string
          description: The unique identifier of the game room.
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        code:
          type: string
      required:
        - roomId
        - code
    GetRoomCodeDto:
      type: object
      properties:
        roomId:
          type: string
          description: >-
            The unique identifier of the game room to which the team will be
            added.
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
      required:
        - roomId
    GetRoomCodeResponseDto:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    CurrentGameResponseDto:
      type: object
      properties:
        roomId:
          type: string
          description: The unique identifier of the game
          example: 123e4567-e89b-12d3-a456-426614174000
      required:
        - roomId
    ValidateCodeDto:
      type: object
      properties:
        code:
          type: string
        roomId:
          type: string
          description: The unique identifier of the game room
          example: 123e4567-e89b-12d3-a456-426614174000
      required:
        - code
        - roomId
    ValidateCodeResponseDto:
      type: object
      properties:
        valid:
          type: boolean
      required:
        - valid
    ParticipantDisplayDataDto:
      type: object
      properties:
        isRegistered:
          type: boolean
        userId:
          type: object
          nullable: true
        name:
          type: string
        avatarUrl:
          type: string
      required:
        - isRegistered
        - userId
        - name
        - avatarUrl
    GameSummaryParticipantDto:
      type: object
      properties:
        participantId:
          type: string
        teamId:
          type: string
        score:
          type: number
        displayData:
          $ref: '#/components/schemas/ParticipantDisplayDataDto'
      required:
        - participantId
        - teamId
        - score
        - displayData
    RoundSummaryDto:
      type: object
      properties:
        id:
          type: string
        roundNumber:
          type: number
        teamId:
          type: string
        guesserParticipantId:
          type: string
      required:
        - id
        - roundNumber
        - teamId
        - guesserParticipantId
    GameSummaryTeamDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    GameSummaryResponseDto:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        createdAt:
          format: date-time
          type: string
        settings:
          $ref: '#/components/schemas/GameSettingsDto'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/GameSummaryParticipantDto'
        roundsSummary:
          type: array
          items:
            $ref: '#/components/schemas/RoundSummaryDto'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/GameSummaryTeamDto'
      required:
        - id
        - status
        - createdAt
        - settings
        - participants
        - roundsSummary
        - teams
    PaginatedGameSummaryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GameSummaryResponseDto'
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
      required:
        - items
        - total
        - limit
        - offset
    WordResponseDto:
      type: object
      properties:
        id:
          type: string
          example: uuid-word-1
          description: Word ID
        text:
          type: string
          example: Apple
          description: The word text
        score:
          type: number
          example: Apple
          description: The word text
      required:
        - id
        - text
        - score
    ParticipantStatsDto:
      type: object
      properties:
        participantId:
          type: string
        teamId:
          type: string
        scoreAfterRound:
          type: number
      required:
        - participantId
        - teamId
        - scoreAfterRound
    RoundDetailsResponseDto:
      type: object
      properties:
        roundId:
          type: string
        number:
          type: number
        words:
          description: Words for the round
          type: array
          items:
            $ref: '#/components/schemas/WordResponseDto'
        participantsStats:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantStatsDto'
      required:
        - roundId
        - number
        - words
        - participantsStats
    PaginatedRoundDetailsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoundDetailsResponseDto'
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
      required:
        - items
        - total
        - limit
        - offset